Ripped out a lot of code around the RA TreeView
This was just proving to be a complete untenable part of the
application. The tree view and its recursive node structure was
imposing another layer on top and was difficult to use. We’re better
off fixing this in a UI. The UX around this would have been complicated
anyways.